update 11/05/2021 By huykhanh

+++ src/main/java/com/dhk/Respository/IDoctorRepository.java
@@ -1,11 +1,13 @@
 package com.dhk.Respository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.dhk.entity.Department;
 import com.dhk.entity.Doctor;
 
 @Repository
-public interface IDoctorRepository extends JpaRepository<Doctor, Long>{
-
+public interface IDoctorRepository extends JpaRepository<Doctor, Integer>{
+	
 }
